refactor(core): move LitharchTag/LitharchState to global namespace
LitharchTag and LitharchState were defined inside TheWaningBorder.Entities
namespace in Litharch.cs entity factory file. Moved them to UnitComponents.cs
in global namespace to match all other unit components (ArcherTag, MinerTag,
etc.). Removed unnecessary using directive from LitharchHealingSystem.
