+++ Assets/Scripts/Entities/Units/Litharch.cs
@@ -120,31 +120,4 @@ namespace TheWaningBorder.Entities
             return entity;
         }
     }
-
-    // ═══════════════════════════════════════════════════════════════════════
-    // LITHARCH COMPONENTS
-    // ═══════════════════════════════════════════════════════════════════════
-
-    /// <summary>
-    /// Marker tag for Litharch healer units.
-    /// </summary>
-    public struct LitharchTag : IComponentData { }
-
-    /// <summary>
-    /// Litharch healer state tracking.
-    /// </summary>
-    public struct LitharchState : IComponentData
-    {
-        /// <summary>Current unit being healed</summary>
-        public Entity HealTarget;
-
-        /// <summary>Time accumulator for healing ticks</summary>
-        public float HealTimer;
-
-        /// <summary>1 if actively healing, 0 otherwise</summary>
-        public byte IsHealing;
-
-        /// <summary>Timer for searching for new heal targets</summary>
-        public float SearchTimer;
-    }
 }
